ai-engine/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Database & Cache
pymongo>=4.6.0
redis>=5.0.0
motor>=3.3.0

# AI/ML Libraries
openai>=1.6.0
transformers>=4.36.0
torch>=2.2.0
scikit-learn>=1.3.0
prophet>=1.1.0
spacy>=3.7.0
nltk>=3.8.0

# OCR & Image Processing
pytesseract>=0.3.10
easyocr>=1.7.0
Pillow>=10.0.0
opencv-python>=4.8.0

# Background Jobs
celery>=5.3.0
flower>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.0
aiofiles>=23.2.0
httpx>=0.25.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
flake8>=6.1.0